littlelemon_drf_mysql/requirements.txt
Django==4.2.14
djangorestframework==3.15.2
mysqlclient==2.2.4
pytz==2024.1